Add listing commands
For legacy purposes, separate command groups were re-added for blacklists and whitelists.
There's a new command group for filters.

Not specifying a list type for the `filter list` command will cause the bot to try to infer whether there's only one kind of list,
for example `!filter list tokens` will pull up the blacklist since there's no whitelist.

If a required field is missing, the user will be prompted to complete it from a selection.

Some level of flexibility was added for the list type in the `filter list` command. For example, a list type can be "DENY", but it will also match "blacklist", "denylist", "deny", "denied", "blacklisted" etc.