Fix argument completion for non-last args
Previously the completed arg would be appended to the end of the message, which didn't work if the missing argument wasn't the last one (for example, a value was given to a following argument because of a failed conversion for previous ones).
The select now employs a different strategy of storing the provided args, and the position where the new are should be inserted.