Add system description commands
The system is fairly complex, and to avoid having to delve into the source code, this commit adds several commands to show descriptions for various components of the system:
- Filters
- Filter lists
- Settings (both from the DB schema and filter-type-specific)

The names that can be described are the ones which are actually being used in the system. So if no tokens list was loaded, it can't be described even if there is an implementation for it.
If no name is relayed to the commands, the list of options will be shown instead.