Remove aio-pika and psutil
aio-pika was used in the RabbitMQ cog (added in https://github.com/python-discord/bot/pull/90) which has since been deleted. It is no longer used.

psutil was used to automatically determine how many threads to use when testing.
However, this package has been quite slow to publish wheels for new Python versions.
Since we don't use it for anything mission critical, I have removed it and hardcoded the number of threads to use when when testing to 8.
An error is not raised it the number of threads used exceeds the number of actual threads available.