Speed Up Sync Cog Loading
The user syncer was blocking the startup of the sync cog due to having
to perform thousands of pointless member fetch requests. This speeds up
that process by increasing the probability that the cache is up-to-date
using `Guild.chunked`, and limiting the fetches to members who were in
the guild during the previous sync only.

Co-authored-by: ChrisJL <chrislovering@users.noreply.github.com>
Co-authored-by: wookie184 <wookie1840@gmail.com>
Signed-off-by: Hassan Abouelela <hassan@hassanamr.com>