Don't decode empty content as json in thread bumper
Closes #2162 Closes BOT-343

self.bot.api_client.request can not be used to check if a thread id exists in the bumped thread list on site, as self.bot.api_client.request() always attempts to decode the response as json.

The thread bump viewset in site returns a 204 if the given thread exists with no content, so this does not work.