Ignore auto_moderation_action messages when applying anti-spam rules
When a message triggers an auto_moderation_action notification, Discord re-writes the author field for the system message to look as if it's from the original author.

This means those messages counted towards our anti-spam filter. This both triggered the filter too early, and also resulted in those auto mod messages being deleted when the filter was hit.