Allow users to eval code in either 3.11 or 3.10
To do this we need to track, for each user, the active eval's code, python version and response message. This is because a button press can now also trigger a job to continue. If we did not track these, then editing your own code and then re-evaluating it would trigger the wait_fors in continue_job for each time you pressed the button to change languages.

Co-authored-by: Mark <1515135+MarkKoz@users.noreply.github.com>