Use interaction.defer for snekbox version switch button
This is so that we do not need to spawn the run_job call in a seperate task.

This also wraps interaction.message.delete() in a NotFound suppress to cover the case where a user re-runs code and very quickly clicks the button. The log arg on send_job will stop the actual job from running in this case.