Correctly initialise redis tests
Calling the cog_load from within the setUp function resulted in interaction with a RedisSession before it was initialised.
This wasn't noticed in CI as it only error under certain concurrency timings due to xdist.

To resolve this, we moved the setup and async setup logic to a base class.

Co-authored-by: Hassan Abouelela <hassan@hassanamr.com>