Ignore mounted in-project venvs on startup
Poetry's virtualenvs.in-project config deafults to None, meaning it will use in-project venvs if it finds one, otherwise it will use the cache dir.
In dev we mount the entire root project directory to /bot. This means if the host's venv in in the project dir, this will get mounted and prioritised by poetry run.
If the host is on a non-linux OS this will cause poetry to fail to boot.