Fix Poetry 1.2 Support
Poetry 1.2 introduced a regression which broke pip `--user` installs.
These types of install were the main way we did installations in docker
and CI, as they made it much more convenient to control the location,
availability, and caching of packages.

Poetry's team does not recognize this as a supported use case, so major
changes were required to get everything working again. Most of the
changes were consolidated into chrislovering/python-poetry-base for
docker, and HassanAbouelela/setup-python for CI.

This is a follow-up to #2268 and #2276.

Signed-off-by: Hassan Abouelela <hassan@hassanamr.com>