Filter adding commands, simplify infraction
This commit adds the mechanic to add new filters.

The overrides of the settings can be provided in the command itself, but also through a UI made of Discord components.
The UI adjusts itself to the data type of the setting, e.g a boolean setting will invoke a select with "True" and "False" options.

This commit additionally gets rid of the mechanic to apply a superstar alongside another higher tier infraction for the same message. This is an edge case that isn't worth the complexity at the moment.

Includes a small fix to the Ping setting which made the __or__ method malfunction.