Bring back enabled categories, remove redundant validators
There needs to be a way to only enable a filter in a specific category, so this setting now fulfills that role. Disabled channels can be used to disable a filter in a specific channel within the category.

Additionally the validators which maybe convert to int are now removed. As long as the int is the first in the Union, it will happen anyway.