domain/exact -> domain/subdomains
The original plan was to have a field which only matches subdomains and not the domain itself, I got confused when I wrote the /exact field.

This fixes a bug where the matches would be updated even if it didn't meet the criteria of the extra field.

Also fixes some issues in the UI dealing with filter overrides.