Change override handling
The Settings class is now a subclass of dict as well.

SettingsEntry's now store which fields are overrides in case the entry is not empty.
If there is a overridden value and a non-overridden value in the same entry, the non-overridden value will now correctly be the default instead of staying empty.

Additionally fixes bugs with free list input parsing.