Add antispam filter list and duplicates filter
Adds the antispam filterlist, which dispatches a set of messages sent in the last X seconds to its filters.
Unlike other filter lists, it doesn't just return the actions to be taken, but rather splits them in two: any actions unrelated to alerting mods are returned as usual, while actions related to alerting are used in a delayed coroutine, from which the alert is sent.

- Fixes `infraction_channel` not using the context channel when it's None.
- Moves the `upload_log` function outside the ModLog cog, as it only needed it for the bot instance.
- Any auto-deleted message with attachments will now have its attachments logged.