Filter match now shows all triggers instead of messages
The purpose of this command is to be able to find any relevant filters. It was not possible to find whitelist filters in the previous method, as there are no alerts for whitelist triggers.

Instead, each filterlist now additionally returns what filters were triggered while processing the context.

Additionally, a `no_user` argument was added to `filter match` to allow checking for triggers on one's own messages (otherwise there's almost no filtering on a mod's messages).