Abbreviate automated messages
This commit abbreviates the automated messages that are triggered by code not posted in markdown blocks, or posted in ill-formatted blocks. Note that I do not have a bot development environment, and if we agree to proceed with this rewording (or a variation thereof), someone will still need to verify that they would be rendered correctly in production.