Add nickname filter
The nickname filter works in much the same way as the one in the old system, with the following changes:
- The lock is per user, rather than a global lock.
- The alert cooldown is one hour, instead of three days which seemed too much.

The delete_messages setting was changed to the more generic remove_context.
If it's a nickname event, the context will be removed by applying a superstar infraction to the user.

In order to allow filtering nicknames in voice state events, the filter context can now have None in the channel field.

Additionally:
- Fixes a bug when ignoring filters in message edits.
- Makes the invites list keep track of message edits.
- The FakeContext class is moved to utils since it's now also needed by remove_context.