Handle Discord API Not Being Ready
When discord sends us the thread create event in help channels, it is
not ready to perform other operations on the thread such as getting or
pinning messages. This causes it to error out when we try to do these
actions and claim that those channels don't exist. Instead, we sleep for
a short time to try and wait for it to be ready.

Signed-off-by: Hassan Abouelela <hassan@hassanamr.com>