Check If Thread Is Closed In wait_for_deletion
The wait_for_deletion utility would try to remove reactions from a
message after the timeout expires, which would normally be fine. In
threads however, they can be closed while waiting for the timeout to
expire. In such a case, the bot will try to remove the reactions after
the channel has been closed and fail. A special exception was added for
this case to do nothing, since this is only a QoL feature.

Signed-off-by: Hassan Abouelela <hassan@hassanamr.com>