Auto archive help forum posts after inactivity
I decided to keep the archive logic simple, and just go for 30 minutes since last message, rather than the hybrid of 30 mins + 10 depending on who sent the last message.
The reason for using the hybrid approach previously was due to us running out of channels frequently
Since this is no longer a problem, I decided to keep the logic simple.