Add alert view
The view provides info about the offending user.
This change required changing the cog to generate a fresh webhook on startup so that it's authenticated to the bot.
Only webhooks authenticated to the bot can send message components.

In the future, this view might be used for more advanced usages such as having a button to pardon the user.