Add warning not to pip install unrelated packages.
This is in response to recent attacks where the attacker DMs someone who opens a thread and tells the thread opener to pip install a malicious package under the false pretense that it will solve their problem. The warning replaces "For more tips, check out our guide on asking good questions", which is stated in the thread closer embed.