Create separate message for warnings without duration, expiry, or "to appeal this infraction ...".
Since warnings involve no change in permissions, the DM that users receive said "Duration: Permanent \\ Expires: Never", which some interpreted to mean that we hold a permanent grudge. This is not intended, so this commit creates a separate template that omits those fields. Some users also thought they could appeal a warning to have it removed from the record, but this is not the case, so this commit creates a separate footer that removes "or appeal" from the instruction to contact the ModMail bot.