Remove help cooldown role
The new forum-based help system implements its own cooldown system which is currently set to restrict members to a new post every 2 minutes.

This restriction in addition to the much higher allowance of threads in our server, means that this help cooldown role is no longer useful.

In fact, it is detrimental. Discord, in its infinite wisdom, tied the ability to change a post's name and tags to the ability to the 'Send messages' permission, meaning users with the help cooldown role were unable to complete these actions.