feat: raw context menu support
Added the context menu object to the command tree
and finished the callback. The only thing iffy
about it is this: due to the nature of
interactions, only one message can be sent as a
response to an interaction, with followup
webhooks needed for further response.
`paginator.pages` is at LEAST 2 elements long,
meaning I can only send either element 0
(the "header") or element 1, which is the ACTUAL
raw content. I opted to only send element 1.
I did try followups in order to send both, but I
felt the result was janky, especially since I
implemented the response to be ephemeral.

Also imported Interaction object in the `from
discord import ...` for easier typehinting.