Rewriting of non-code sections.
The previous version began with "Iterating over range-len is a common approach ...", which sounds like an endorsement. Also removed "is guaranteed to produce elements in the same order", as I don't think anyone is actually confused about that. (I wrote the previous version--this is a criticism of my own writing.)