Branding: fail if no event is active
If the bot cannot find an active event, it will try to find the
fallback event. If the fallback event cannot be found either,
it will now raise an error.

This should be an exceptional state that cannot be recovered from.
Synchronisation should be retried once the event setup is corrected.