Remove the help-dm feature
This feature allowed users to opt-in to getting DMs from the bot whenever someone messaged in a help channel they we participating in. This was useful in the old channel-based help system as it was hard to remember which channels you were participating in.

Now that we've moved over to the forum channel, this is no longer an issue, as posts you are participating in now appear in the channel list, and others don't.