Allow for more than 3 nomination votes if others are ticketed
This changes how we calculate whether we are ready for another vote in
the talentpool autonomination system.

Instead of hard-locking at 3 votes in the pool, we factor in whether an
admin-task has been created for the vote (signified by a reaction with
the :ticket: emoji).

We allow for up to 10 votes to be in the pool total, which in theory
allows 3 active votes and 7 votes waiting on the admin-task
completion (though we should never hit this threshold realistically).

This should allow for voting to be more or less continuous and ensure
there is always new votes in the pool and we are not blocked on the
completion of a Helpering/waiting for a Helper candidate to be available
for a new vote to start.

The thresholds are based on those suggested in the nomination-discussion
moderator channel.