Post nomination entries into the voting thread instead of the root
This change moves us from posting multiple messages in the
nomination-voting channel to instead posting all entries into the voting
thread.

This makes the nomination process guaranteed to be contained within one
root message and one channel, instead of splitting into multiple
messages in the root if we went over the character limit.

It greatly simplifies both the creation and the archival of nomination
votes as we only need to handle one message.

It also means that all nominations are guaranteed to be visible in the
voting thread.